# api-server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.terminateSessionRequest:
    properties:
      cooldown_minutes:
        type: integer
    type: object
...
  services.Alert:
    properties:
      count:
        type: integer
      kind:
        type: string
      severity:
        type: string
    type: object
  services.Alerts:
    properties:
      evaluated_at:
        type: string
      items:
        items:
          $ref: '#/definitions/services.Alert'
        type: array
      node_issues:
        items:
          $ref: '#/definitions/services.NodeIssue'
        type: array
      total:
        description: |-
          Total counts only what an operator still has to look at: firing conditions
          that are neither acknowledged nor silenced. Stale rows are excluded because
          their readings are frozen.
        type: integer
    type: object
  services.Entry:
    properties:
      actor_id:
        type: string
      actor_label:
        type: string
      actor_type:
        type: string
      created_at:
        type: string
      detail:
        additionalProperties: {}
        type: object
      event_type:
        type: string
      id:
        type: string
      severity:
        type: string
      target_id:
        type: string
      target_type:
        type: string
    type: object
...
  services.NodeIssue:
    properties:
      acked:
        type: boolean
      acked_by:
        type: string
      alert_id:
        description: |-
          Lifecycle fields, added when alerts became stateful. Additive on purpose:
          an older client that ignores unknown keys still reads the same response.
        type: string
      country:
        type: string
      duration_seconds:
        type: integer
      fired_at:
        type: string
      kind:
        type: string
      node_deleted:
        description: |-
          NodeDeleted marks an alert whose node has been removed. Such an alert can
          never be resolved by evaluation - nothing reports for it - so the UI offers
          closing it by hand.
        type: boolean
      node_id:
        type: string
      node_name:
        type: string
      region:
        type: string
      severity:
        type: string
      silenced_until:
        type: string
      state:
        type: string
      status:
        type: string
      value:
        type: number
    type: object
  services.NodeTraffic:
    properties:
      download:
        type: integer
      node_id:
        type: string
      node_name:
        type: string
      status:
        type: string
      upload:
        type: integer
    type: object
...
host: localhost:2053
info:
  contact: {}
  description: VPN Panel Management API
  title: Proxima VPN API
  version: 0.1.0
paths:
  /admin/activity:
    get:
      description: Returns the most recent recorded panel events, newest first. Supplying
        both target_type and target_id narrows the feed to that target.
      parameters:
      - description: Maximum entries to return (default 20)
        in: query
        name: limit
        type: integer
      - description: Restrict to one target type, e.g. node (requires target_id)
        in: query
        name: target_type
        type: string
      - description: Restrict to one target id (requires target_type)
        in: query
        name: target_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.Entry'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get recent activity
      tags:
      - admin-stats
...
  /admin/devices/{id}/terminate:
    post:
      consumes:
      - application/json
      description: Evicts a device so the node agent withdraws its credential on the
        next config poll
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      - description: Cooldown in minutes
        in: body
        name: body
        schema:
          $ref: '#/definitions/handlers.terminateSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Terminate a device's session
      tags:
      - admin-stats
...
  /admin/online-users:
    get:
      description: Returns detailed info about currently connected users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties:
                type: string
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get online users
      tags:
      - admin-stats
...
  /admin/stats:
    get:
      description: Returns aggregate statistics for the admin dashboard
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get dashboard stats
      tags:
      - admin-stats
  /admin/stats/alerts:
    get:
      description: Returns offline nodes and resource-pressured nodes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.Alerts'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get dashboard alerts
      tags:
      - admin-stats
  /admin/stats/node-traffic:
    get:
      description: Returns upload/download per node over the requested window
      parameters:
      - description: today, week, or month (default today)
        in: query
        name: window
        type: string
      - description: Maximum nodes to return (default 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.NodeTraffic'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get per-node traffic
      tags:
      - admin-stats
  /admin/stats/traffic-history:
    get:
      description: Returns daily upload/download traffic aggregated from traffic_logs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get traffic history
      tags:
      - admin-stats
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"